+++ changelog.rst
@@ -1,3 +1,3 @@
-* 11/09/15: changed the databaseworker in order to save and restore node_address set for the crashes with pickle
+* 11/09/15: changed the databaseworker in order to save and restore the node_address sets for the crashes with pickle
 * 10/25/15: now the Python SimpleHTTPServer module will be started in the testcase directory if the program option use_http is set
 * 10/24/15: implemented the possibility to feed multiple programs with the same input data on one node